+++ demisto_sdk/commands/content_graph/content_graph_commands.py
@@ -59,10 +59,11 @@ def update_content_graph(
     if packs_to_update is None:
         packs_to_update = []
     builder = ContentGraphBuilder(content_graph_interface)
-    if not imported_path and not use_current:
-        # getting the graph from remote, so we need to clean the import dir
+    if not use_current:
         content_graph_interface.clean_import_dir()
-        extract_remote_import_files(content_graph_interface, builder)
+        if not imported_path:
+            # getting the graph from remote, so we need to clean the import dir
+            extract_remote_import_files(content_graph_interface, builder)
 
     if use_git and (commit := content_graph_interface.commit):
         packs_to_update.extend(GitUtil().get_all_changed_pack_ids(commit))