+++ demisto_sdk/commands/content_graph/interface/neo4j/neo4j_graph.py
@@ -1,7 +1,7 @@
 import logging
 from multiprocessing import Pool
 from pathlib import Path
-from typing import Any, Dict, Iterable, List, Optional, Set
+from typing import Any, Dict, Iterable, List, Optional, Set, Tuple
 
 from neo4j import GraphDatabase, Neo4jDriver, Session, graph
 
@@ -47,7 +47,6 @@ from demisto_sdk.commands.content_graph.interface.neo4j.queries.nodes import (
     _match,
     create_nodes,
     delete_all_graph_nodes,
-    duplicates_exist,
     get_relationships_to_preserve,
     remove_empty_properties,
     remove_packs_before_creation,
@@ -57,6 +56,14 @@ from demisto_sdk.commands.content_graph.interface.neo4j.queries.nodes import (
 from demisto_sdk.commands.content_graph.interface.neo4j.queries.relationships import (
     _match_relationships,
     create_relationships,
+)
+from demisto_sdk.commands.content_graph.interface.neo4j.queries.validations import (
+    validate_dependencies,
+    validate_fromversion,
+    validate_marketplaces,
+    validate_multiple_packs_with_same_display_name,
+    validate_toversion,
+    validate_unknown_content,
 )
 from demisto_sdk.commands.content_graph.objects.base_content import (
     BaseContent,
@@ -116,7 +123,7 @@ class Neo4jContentGraphInterface(ContentGraphInterface):
             auth=(NEO4J_USERNAME, NEO4J_PASSWORD),
         )
         if should_update:
-            update_content_graph(self)
+            update_content_graph(self, use_git=True)
 
     def __enter__(self) -> "Neo4jContentGraphInterface":
         return self
@@ -141,7 +148,7 @@ class Neo4jContentGraphInterface(ContentGraphInterface):
         self,
         session: Session,
         result: Dict[int, Neo4jRelationshipResult],
-        marketplace: Optional[MarketplaceVersions],
+        marketplace: Optional[MarketplaceVersions] = None,
     ):
         """This adds relationships to given object
 
@@ -251,7 +258,7 @@ class Neo4jContentGraphInterface(ContentGraphInterface):
                     ),
                 )
 
-    def _add_nodes_to_mapping(self, nodes: List[graph.Node]) -> None:
+    def _add_nodes_to_mapping(self, nodes: Iterable[graph.Node]) -> None:
         """Add nodes to the content models mapping
 
         Args:
@@ -327,10 +334,103 @@ class Neo4jContentGraphInterface(ContentGraphInterface):
             session.write_transaction(create_nodes, nodes)
             session.write_transaction(remove_empty_properties)
 
-    def validate_graph(self) -> None:
+    def get_unknown_content_uses(self, file_paths: List[str]) -> List[BaseContent]:
         with self.driver.session() as session:
-            if session.read_transaction(duplicates_exist):
-                raise Exception("Duplicates found in graph.")
+            results: Dict[int, Neo4jRelationshipResult] = session.read_transaction(
+                validate_unknown_content, file_paths
+            )
+            self._add_nodes_to_mapping(result.node_from for result in results.values())
+            self._add_relationships_to_objects(session, results)
+            return [self._id_to_obj[result] for result in results]
+
+    def get_duplicate_pack_display_name(
+        self, file_paths: List[str]
+    ) -> List[Tuple[str, List[str]]]:
+        with self.driver.session() as session:
+            results = session.read_transaction(
+                validate_multiple_packs_with_same_display_name, file_paths
+            )
+            return results
+
+    def find_uses_paths_with_invalid_fromversion(
+        self, file_paths: List[str], for_supported_versions=False
+    ) -> List[BaseContent]:
+        """Searches and retrievs content items who use content items with a lower fromvesion.
+
+        Args:
+            file_paths (List[str]): A list of content items' paths to check.
+                If not given, runs the query over all content items.
+
+        Returns:
+            List[BaseContent]: The content items who use content items with a lower fromvesion.
+        """
+        with self.driver.session() as session:
+            results: Dict[int, Neo4jRelationshipResult] = session.read_transaction(
+                validate_fromversion, file_paths, for_supported_versions
+            )
+            self._add_nodes_to_mapping(result.node_from for result in results.values())
+            self._add_relationships_to_objects(session, results)
+            return [self._id_to_obj[result] for result in results]
+
+    def find_uses_paths_with_invalid_toversion(
+        self, file_paths: List[str], for_supported_versions=False
+    ) -> List[BaseContent]:
+        """Searches and retrievs content items who use content items with a higher toversion.
+
+        Args:
+            file_paths (List[str]): A list of content items' paths to check.
+                If not given, runs the query over all content items.
+
+        Returns:
+            List[BaseContent]: The content items who use content items with a higher toversion.
+        """
+        with self.driver.session() as session:
+            results: Dict[int, Neo4jRelationshipResult] = session.read_transaction(
+                validate_toversion, file_paths, for_supported_versions
+            )
+            self._add_nodes_to_mapping(result.node_from for result in results.values())
+            self._add_relationships_to_objects(session, results)
+            return [self._id_to_obj[result] for result in results]
+
+    def find_uses_paths_with_invalid_marketplaces(
+        self, file_paths: List[str]
+    ) -> List[BaseContent]:
+        """Searches and retrievs content items who use content items with invalid marketplaces.
+
+        Args:
+            file_paths (List[str]): A list of content items' paths to check.
+                If not given, runs the query over all content items.
+
+        Returns:
+            List[BaseContent]: The content items who use content items with invalid marketplaces.
+        """
+        with self.driver.session() as session:
+            results: Dict[int, Neo4jRelationshipResult] = session.read_transaction(
+                validate_marketplaces, file_paths
+            )
+            self._add_nodes_to_mapping(result.node_from for result in results.values())
+            self._add_relationships_to_objects(session, results)
+            return [self._id_to_obj[result] for result in results]
+
+    def find_core_packs_depend_on_non_core_packs(
+        self, pack_ids: List[str], core_pack_list
+    ) -> List[BaseContent]:
+        """Searches and retrieves core packs who depends on content items who are not core packs.
+
+        Args:
+            pack_ids (List[str]): A list of content items pack_ids to check.
+            core_pack_list: A list of core packs
+
+        Returns:
+            List[BaseContent]: The core packs who depends on content items who are not core packs.
+        """
+        with self.driver.session() as session:
+            results: Dict[int, Neo4jRelationshipResult] = session.read_transaction(
+                validate_dependencies, pack_ids, core_pack_list
+            )
+            self._add_nodes_to_mapping(result.node_from for result in results.values())
+            self._add_relationships_to_objects(session, results)
+            return [self._id_to_obj[result] for result in results]
 
     def create_relationships(
         self, relationships: Dict[RelationshipType, List[Dict[str, Any]]]