+++ demisto_sdk/commands/content_graph/interface/neo4j/queries/common.py
@@ -5,6 +5,7 @@ from pathlib import Path
 from typing import Any, Dict
 
 from neo4j import Result, Transaction
+from packaging.version import Version
 
 from demisto_sdk.commands.content_graph.common import ContentType
 
@@ -16,6 +17,11 @@ def labels_of(content_type: ContentType) -> str:
 
 
 def versioned(property: str) -> str:
+    try:
+        Version(property)
+        property = f'"{property}"'
+    except Exception:
+        pass
     return f'toIntegerList(split({property}, "."))'