+++ demisto_sdk/commands/content_graph/interface/neo4j/queries/import_export.py
@@ -21,6 +21,7 @@ LIST_PROPERTIES = [
     "enhancement_script_names",
     "definition_ids",
     "tabs",
+    "excluded_dependencies",
 ]
 
 CONVERT_FIELD_TO_STRING = """MATCH (n) WHERE NOT n.{prop} IS NULL