+++ demisto_sdk/commands/content_graph/tests/create_content_graph_test.py
@@ -72,6 +72,7 @@ def mock_pack(name: str = "SamplePack"):
         useCases=[],
         keywords=[],
         contentItems=[],
+        excluded_dependencies=[],
     )