+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 ## Unreleased
 * **Breaking change**: Removed usage of `pipenv`, `isort` and `autopep8` in the **split** and **download** commands. Removed the `--no-pipenv` flag. Please see https://xsoar.pan.dev/docs/tutorials/tut-setup-dev-remote for the recommended environment setup.
 * Fixed an issue in **prepare-content** command where large code lines were broken.
@@ -12,6 +13,8 @@
 * Added support for the **Layout Rule** content type in the id-set and the content graph.
 * Updated the default general `fromVersion` value on **format** to `6.8.0`
 * Fixed an issue where **lint** sometimes failed when using the `-cdam` flag due to wrong file duplications filtering.
+* Updated **validate** to use the graph when running with `--graph` flag.
+
 
 ## 1.9.0
 * Fixed an issue where the Slack notifier was using a deprecated argument.